requirements.txt
urllib3
pillow
imagehash
piexif